+++ Source/Creation/Frame/Content/String/Encoding/ID3UTF16StringToByteAdapter.swift
@@ -1,0 +1,23 @@
+//
+//  ID3UTF16StringToByteAdapter.swift
+//
+//  Created by Fabrizio Duroni on 27/04/2018.
+//  2018 Fabrizio Duroni.
+//
+
+import Foundation
+
+class ID3UTF16StringToByteAdapter: StringToBytesAdapter {
+    private let paddingAdder: PaddingAdder
+    private let frameConfiguration: ID3FrameConfiguration
+    
+    init(paddingAdder: PaddingAdder, frameConfiguration: ID3FrameConfiguration) {
+        self.paddingAdder = paddingAdder
+        self.frameConfiguration = frameConfiguration
+    }
+    
+    func adapt(string: String, for version: ID3Version) -> [UInt8] {
+        return frameConfiguration.encodingByteFor(version: version, encoding: .UTF16) +
+            paddingAdder.addTo(content: [UInt8](string.data(using: .utf16) ?? Data()), numberOfByte: 2)
+    }
+}