+++ Source/Frame/ID3FrameConfiguration.swift
@@ -64,6 +64,21 @@ class ID3FrameConfiguration {
             "TRCK" : .TrackPosition
         ]
     ]
+    private let encodingPositionInBytes: [ID3Version : Int] = [
+        .version2 : 6,
+        .version3 : 10
+    ]
+    private let encodingSizeInBytes: Int = 1
+    private let encodingByte: [ID3Version : [ID3StringEncoding : [UInt8]]] = [
+        .version2 : [
+            .ISO88591 : [0x00],
+            .UTF16 : [0x01],
+        ],
+        .version3 : [
+            .ISO88591 : [0x00],
+            .UTF16 : [0x01],
+        ]
+    ]
 
     func headerSizeFor(version: ID3Version) -> Int {
         return headerSizesInBytes[version]!
@@ -87,5 +102,17 @@ class ID3FrameConfiguration {
 
     func frameTypeFor(identifier: String, version: ID3Version) -> FrameType {
         return nameForIdentifier[version]![identifier] ?? .Invalid
+    }
+    
+    func encodingSize() -> Int {
+        return encodingSizeInBytes
+    }
+    
+    func encodingPositionFor(version: ID3Version) -> Int {
+        return encodingPositionInBytes[version]!
+    }
+
+    func encodingByteFor(version: ID3Version, encoding: ID3StringEncoding) -> [UInt8] {
+        return encodingByte[version]![encoding]!
     }
 }