+++ Source/Padding/Add/PaddingAdder.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 protocol PaddingAdder {
-    func addTo(content: [UInt8]) -> [UInt8]
+    func addTo(content: [UInt8], numberOfByte: Int) -> [UInt8]
 }