+++ Source/Parsing/Frame/Content/Operation/ID3StringEncoding.swift
@@ -1,0 +1,13 @@
+//
+//  ID3StringEncoding.swift
+//  ID3TagEditor
+//
+//  Created by Fabrizio Duroni on 27/04/18.
+//
+
+import Foundation
+
+enum ID3StringEncoding: UInt8 {
+    case ISO88591
+    case UTF16
+}