+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All changes to ID3TagEditor will be documented in this file.
 
+## [1.1.1](https://github.com/chicio/ID3TagEditor/releases/tag/1.1.1)
+Release date: 2018-04-29.
+
+#### Added
+- `Genre`, `AttachedPicture` and `TrackPositionInSet` implement `Equatable` protocol.
+
+#### Fixed
+- Fixed encoding for 'text information' frames: now strings in this frame are encoding using utf16 string (with BOM).
+
+  
 ## [1.0.0](https://github.com/chicio/ID3TagEditor/releases/tag/1.0.0)
 Release date: 2018-03-14.